pyproject.toml
[tool.poetry]
name = "orcid-link"
version = "0.1.0"
description = ""
authors = ["Erik Pearson <eapearson@lbl.gov>"]

[tool.poetry.dependencies]
python = "3.11.0"
dicttoxml2 = "2.1.0"
fastapi = "0.88.0"
uvicorn = "0.20.0"
xmltodict = "0.13.0"
pymongo = "4.3.3"
PyYAML = "6.0"
httpx = "0.23.3"
pytest = "7.2.0"
pytest-cov = "4.0.0"
pyfakefs = "5.0.0"
cache3 = "0.3.1"


[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = { version = "^22.10.0", allow-prereleases = true }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
